chore: remove substring file
